requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
execnet==2.1.1
fastapi==0.128.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
jiter==0.8.2
lxml==5.3.0
numpy==2.2.2
openai==1.61.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==12.0.0
playwright==1.50.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.11
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PySocks==1.7.1
pytest==9.0.0
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.46
starlette==0.50.0
tk==0.1.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.2
